+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+0.2.0-alpha (Jul 1, 2015)
+--------------------
+ New features:
+  - Server socket support
+  - Support all transports returned by `stream_get_transports`
+  - Distinguish frame boundaries
+  - Determine datagram size for udp sockets
+ 
 0.1.1 (May 25, 2015)
 --------------------
  - Added additional checks due to https://bugs.php.net/bug.php?id=64803